feat(core): remove oidc schemas constants and types
