+++ packages/core/src/metadata-policy/index.ts
@@ -1,8 +1,3 @@
-export {
-	applyMetadataPolicy,
-	denormalizeScope,
-	type MetadataPolicyOptions,
-	normalizeScope,
-} from "./apply.js";
+export { applyMetadataPolicy, type MetadataPolicyOptions } from "./apply.js";
 export { validateCustomOperators } from "./custom-operators.js";
 export { resolveMetadataPolicy } from "./merge.js";
